[CL-2916] Add mock
