+++ front/app/api/idea_statuses/useIdeaStatus.test.ts
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 
 import useIdeaStatus from './useIdeaStatus';
-import { ideaStatusesData } from './useIdeaStatuses.test';
+import { ideaStatusesData } from './__mocks__/useIdeaStatuses';
 
 import { setupServer } from 'msw/node';
 import { rest } from 'msw';
