+++ front/app/api/idea_statuses/useIdeaStatuses.test.ts
@@ -6,39 +6,9 @@ import { setupServer } from 'msw/node';
 import { rest } from 'msw';
 
 import createQueryClientWrapper from 'utils/testUtils/queryClientWrapper';
+import { ideaStatusesData } from './__mocks__/useIdeaStatuses';
 
 const apiPath = '*idea_statuses';
-
-export const ideaStatusesData = [
-  {
-    id: '1',
-    type: 'idea_status',
-    attributes: {
-      code: 'new',
-      name_multiloc: {
-        en: 'New',
-      },
-      color: '#FF0000',
-      ordering: 1,
-      created_at: '2021-03-03T09:00:00.000Z',
-      updated_at: '2021-03-03T09:00:00.000Z',
-    },
-  },
-  {
-    id: '2',
-    type: 'idea_status',
-    attributes: {
-      code: 'in_progress',
-      name_multiloc: {
-        en: 'In progress',
-      },
-      color: '#00FF00',
-      ordering: 2,
-      created_at: '2021-03-03T09:00:00.000Z',
-      updated_at: '2021-03-03T09:00:00.000Z',
-    },
-  },
-];
 
 const server = setupServer(
   rest.get(apiPath, (_req, res, ctx) => {
